+++ popkorn-compiler/src/main/kotlin/cc/popkorn/compiler/generators/MappingGenerator.kt
@@ -1,8 +1,9 @@
 package cc.popkorn.compiler.generators
 
-import cc.popkorn.compiler.utils.PROVIDER_SUFFIX
-import cc.popkorn.compiler.utils.RESOLVER_SUFFIX
-import mapping.Mapping
+import cc.popkorn.PROVIDER_SUFFIX
+import cc.popkorn.RESOLVER_SUFFIX
+import cc.popkorn.compiler.utils.splitPackage
+import cc.popkorn.mapping.Mapping
 import com.squareup.kotlinpoet.*
 import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
 import java.io.File
@@ -18,16 +19,16 @@ import kotlin.reflect.KClass
  */
 internal class MappingGenerator(private val directory: File) {
 
-    fun writeResolvers(moduleName:String, list:Set<TypeMirror>) {
+    fun writeResolvers(className:String, list:Set<TypeMirror>) {
         val code = getCode(list, RESOLVER_SUFFIX)
-        val file = getFile(code, "$moduleName$RESOLVER_SUFFIX")
-        file.writeTo(directory)
+        val file = getFile(code, className)
+        return file.writeTo(directory)
     }
 
 
-    fun writeProviders(moduleName:String, list:List<TypeElement>) {
+    fun writeProviders(className:String, list:List<TypeElement>) {
         val code = getCode(list.map { it.asType() }.toSet(), PROVIDER_SUFFIX)
-        val file = getFile(code, "$moduleName$PROVIDER_SUFFIX")
+        val file = getFile(code, className)
         file.writeTo(directory)
     }
 
@@ -36,25 +37,27 @@ internal class MappingGenerator(private val directory: File) {
         val function = CodeBlock.builder()
         function.add("return when(original){\n")
         list.forEach {
-            function.add("    $it::class -> ${it}_$suffix::class\n")
+            function.add("    $it::class -> ${it}_$suffix()\n")
         }
         function.add("    else -> null\n")
         function.add("}\n")
         return function.build()
     }
 
-    private fun getFile(creationCode:CodeBlock, filename:String) : FileSpec {
+    private fun getFile(creationCode:CodeBlock, qualifiedName:String) : FileSpec {
+        val producerOfAny = WildcardTypeName.producerOf(Any::class)
         val create = FunSpec.builder("find")
-            .addParameter("original", KClass::class.asClassName().parameterizedBy(STAR))
+            .addParameter("original", KClass::class.asClassName().parameterizedBy(producerOfAny))
             .addModifiers(KModifier.OVERRIDE)
-            .returns(KClass::class.asClassName().parameterizedBy(STAR).copy(nullable = true))
+            .returns(Any::class.asClassName().copy(nullable = true))
             .addCode(creationCode)
             .build()
 
-        val name = "${filename}Mapping"
-        return FileSpec.builder("cc.popkorn.mapping", name)
+
+        val pack = qualifiedName.splitPackage()
+        return FileSpec.builder(pack.first, pack.second)
             .addType(
-                TypeSpec.classBuilder(name)
+                TypeSpec.classBuilder(pack.second)
                     .addSuperinterface(Mapping::class)
                     .addFunction(create)
                     .build()