+++ popkorn-compiler/src/main/kotlin/cc/popkorn/compiler/utils/Constants.kt
@@ -1,5 +1,0 @@
-package cc.popkorn.compiler.utils
-
-
-const val PROVIDER_SUFFIX = "Provider"
-const val RESOLVER_SUFFIX = "Resolver"