+++ popkorn/src/main/kotlin/cc/popkorn/mapping/Mapping.kt
@@ -1,16 +1,16 @@
-package mapping
+package cc.popkorn.mapping
 
 import kotlin.reflect.KClass
 
 /**
- * Interface that creates gets the mapping of a certain class
+ * Interface that creates gets the cc.popkorn.mapping of a certain class
  *
  * @author Pau Corbella
  * @since 1.0
  */
 interface Mapping {
 
-     fun find(original:KClass<*>) : KClass<*>?
+     fun find(original:KClass<out Any>) : Any?
 
      fun isPresent(original:KClass<*>) : Boolean{
           return find(original) != null