+++ popkorn/src/main/kotlin/cc/popkorn/pools/ProviderPool.kt
@@ -2,7 +2,7 @@
 package cc.popkorn.pools
 
 import cc.popkorn.core.Provider
-import mapping.Mapping
+import cc.popkorn.mapping.Mapping
 import kotlin.reflect.KClass
 
 /**
@@ -19,15 +19,14 @@ internal class ProviderPool(private val mappings:LinkedHashSet<Mapping> = linked
 
     fun <T : Any> create(clazz: KClass<T>): Provider<T> {
         return findProvider(clazz)
-            ?.newInstance()
             ?.let { it as? Provider<T> }
             ?: throw RuntimeException("Could not find Provider for this class: ${clazz.qualifiedName}. Did you forget to add @Injectable?")
     }
 
 
-    private fun findProvider(original:KClass<*>) : Class<*>?{
+    private fun findProvider(original:KClass<*>) : Any?{
         mappings.forEach { map ->
-            map.find(original)?.also { return it.java }
+            map.find(original)?.also { return it }
         }
         return null
     }